// warps/multiversx/esdt-nft-view/warp.types.ts
// This file is auto-generated by scripts/generate-warp-types.js

export type MultiversXViewNFTInputs = {
  NFT_IDENTIFIER: string;
};

export type MultiversXViewNFTData = {
  NFT_IDENTIFIER: string;
  NFT_NAME: string;
  COLLECTION: string;
  NONCE: number;
  ROYALTIES: number;
  MEDIA_URL: string;
  THUMBNAIL_URL: string;
  ATTRIBUTES: string;
  CREATOR: string;
  OWNER: string;
};